# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WoT Blitz Stats API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments:
    get:
      operationId: listTournaments
      tags: [tournaments]
      summary: List all tournaments
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [upcoming, active, completed]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of tournaments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentListResponse"
    post:
      operationId: createTournament
      tags: [tournaments]
      summary: Create a new tournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTournamentInput"
      responses:
        "201":
          description: Tournament created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
  /tournaments/{tournamentId}:
    get:
      operationId: getTournament
      tags: [tournaments]
      summary: Get a specific tournament
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"
        "404":
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    TournamentParticipant:
      type: object
      properties:
        account_id:
          type: integer
        nickname:
          type: string
        score:
          type: integer
        rank:
          type: integer
      required:
        - account_id
        - nickname
        - score
        - rank
    Tournament:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [upcoming, active, completed]
        start_date:
          type: string
        end_date:
          type: string
        prize_pool:
          type: string
          nullable: true
        max_participants:
          type: integer
          nullable: true
        current_participants:
          type: integer
        format:
          type: string
        tier_requirement:
          type: integer
          nullable: true
        participants:
          type: array
          items:
            $ref: "#/components/schemas/TournamentParticipant"
          nullable: true
        created_at:
          type: string
      required:
        - id
        - title
        - description
        - status
        - start_date
        - end_date
        - current_participants
        - format
        - created_at
    TournamentListResponse:
      type: object
      properties:
        tournaments:
          type: array
          items:
            $ref: "#/components/schemas/Tournament"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - tournaments
        - total
        - page
        - limit
    CreateTournamentInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        start_date:
          type: string
        end_date:
          type: string
        format:
          type: string
        prize_pool:
          type: string
          nullable: true
        max_participants:
          type: integer
          nullable: true
        tier_requirement:
          type: integer
          nullable: true
      required:
        - title
        - description
        - start_date
        - end_date
        - format
...